move intiial condition generation into separate package velocityfunction
